requirements.txt
amqp==2.2.2
Babel==2.5.1
backports-abc==0.5
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
#Django==1.11
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-crispy-forms==1.7.0
django-debug-toolbar==1.8
django-redis==4.8.0
django-redis-sessions==0.6.1
flower==0.9.2
futures==3.1.1
kombu==4.1.0
pytz==2017.2
redis==2.10.6
singledispatch==3.4.0.3
six==1.11.0
sqlparse==0.2.4
tornado==4.5.2
vine==1.1.4
